+++ restfmclient/__init__.py
@@ -3,12 +3,14 @@
 from restfmclient.client import Client
 from restfmclient.exceptions import RESTfmException
 from restfmclient.exceptions import RESTfmNotFound
+from restfmclient.record import Record
 from restfmclient.version import __version__
 
 
 __all__ = [
     '__version__',
     'Client',
+    'Record',
     'RESTfmException',
     'RESTfmNotFound'
 ]